Pre-release version
This is the pre-release version, things **should** work, but I can't
guarantee as I haven't tested it.

It says `0.1.0` on `Cargo.toml`, that is very inaccurate, but I can't be
bothered to fix it, so when I fully release the app I'll make it `1.0.0`
and then start counting from there.